requirements.txt
# Web Server & Routing
Flask>=3.0.0
Werkzeug>=3.0.0

# Document Parsing (from core/document_parser.py)
PyPDF2>=3.0.0
python-docx>=1.1.0
PyMuPDF>=1.23.0  # This is the 'fitz' library used for best PDF extraction

# LLM & Environment variables (from core/llm_engine.py & document_parser.py)
google-genai>=0.3.0
python-dotenv>=1.0.0

# Machine Learning & Vision (from your original requirements & core/handwriting_engine.py)
gradio
qwen-vl-utils
opencv-python-headless  # Headless is safer for server environments
numpy>=1.26
Pillow>=10.0
Requests>=2.31
torch
torchvision
transformers>=4.45.0
accelerate